basic context switching
Currently, there's only one task, so the "scheduler" simply always
selects task1 as the next task to run. There are still some quirks to
iron out. For example, if task1 looks like this

for( ;; ) {
    if( !(ticks_since_boot % 100) ) {
        PORTB |= _BV( PIN5 );
    } else {
        PORTB &= ~_BV( PIN5 );
    }
}

The tick ISR will stop firing as soon as task1 is first run. This is
because modding by 100 requires disabling, then re-enabling interrupts
behind the scenes. Since it taks at least one cycle for re-enabling
interrupts to take effect, and we're in a loop, the ISR doesn't get a
chance to run. That's the current theory anyway...