add serial_parser.py
This python script reads binary from stdin, and outputs a stream of
packets as human-readible text to stdout.

Note that the parser() function is a generator which takes a byte
reading function as an argument. You could use it, for example, with
pySerial's Serial.read function.

Also note that packet_is_complete() is defined inside parser(). This
both limits it's scope, and allows access to parser's local variables.